removed tests and userService; adapted and renamed userServiceIml
