+++ src/main/java/by/pilipuk/PersonApp.java
@@ -1,4 +1,4 @@
-package jdbc.java;
+package by.pilipuk;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
