+++ src/main/resources/application.yaml
@@ -7,7 +7,6 @@ spring:
 
   jpa:
     hibernate:
-      ddl-auto: update
       show-sql: true
     properties:
       hibernate:
